pyproject.toml
[project]
name = "scraper-redes-sociais"
version = "0.1.0"
description = "Scraper simples para buscar notícias do G1"
requires-python = ">=3.8"
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "psycopg2-binary>=2.9.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

